authen, author by JWT

+++ src/main/java/com/finalproject/StayFinderApi/service/impl/CommentServiceImpl.java
@@ -6,20 +6,24 @@ import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Service;
-
 import com.finalproject.StayFinderApi.dto.AccountRespone;
 import com.finalproject.StayFinderApi.dto.CommentRequest;
 import com.finalproject.StayFinderApi.dto.CommentResponse;
 import com.finalproject.StayFinderApi.entity.Account;
 import com.finalproject.StayFinderApi.entity.Comment;
+import com.finalproject.StayFinderApi.entity.PositionNameEnum;
 import com.finalproject.StayFinderApi.entity.Post;
 import com.finalproject.StayFinderApi.exception.AppException;
+import com.finalproject.StayFinderApi.exception.NotFoundException;
+import com.finalproject.StayFinderApi.exception.StayFinderApiException;
 import com.finalproject.StayFinderApi.repository.AccountRepository;
 import com.finalproject.StayFinderApi.repository.CommentRepository;
 import com.finalproject.StayFinderApi.repository.PostRepository;
+import com.finalproject.StayFinderApi.security.UserPrincipal;
 import com.finalproject.StayFinderApi.service.ICommentService;
 
 @Service
@@ -56,15 +60,22 @@ public class CommentServiceImpl implements ICommentService {
 	}
 
 	@Override
-	public boolean deleteCommentById(long id) {
-		commentRepo.deleteById(id);
-		return true;
+	public boolean deleteCommentById(long id, UserPrincipal userPrincipal) {
+		Comment comment = commentRepo.findById(id).orElseThrow(()-> new NotFoundException("Comment id không tồn tại"));
+		if(comment.getAccount().getId() == userPrincipal.getId() || userPrincipal.getAuthorities().contains(new SimpleGrantedAuthority(PositionNameEnum.ROLE_ADMIN.toString())) ) {
+			commentRepo.deleteById(id);
+			return true;
+		}
+		else 
+			throw new StayFinderApiException(HttpStatus.UNAUTHORIZED, "You don't have permission to delete this comment");
+		
+		
 	}
 
 	@Override
-	public CommentResponse addComment(CommentRequest commentRequest) {
+	public CommentResponse addComment(CommentRequest commentRequest, UserPrincipal userPrincipal) {
 		Optional<Post> postOptional = postRepo.findById(commentRequest.getPostId());
-		Optional<Account> accountOptional = accountRepo.findByUsername(commentRequest.getUsername());
+		Optional<Account> accountOptional = accountRepo.findByUsername(userPrincipal.getUsername());
 		if(postOptional.isPresent() && accountOptional.isPresent())
 		{
 			String imgUrl = commentRequest.getImageUrl() == null ? null : commentRequest.getImageUrl();
@@ -72,7 +83,8 @@ public class CommentServiceImpl implements ICommentService {
 			
 			return new CommentResponse(comment.getId(), commentRequest.getPostId(),new AccountRespone(accountOptional.get().getUsername(), accountOptional.get().getName(), accountOptional.get().getAvatarUrl()) , comment.getContent(), comment.getCommentTime(), comment.getImageUrl());
 		}
-		throw new AppException("Sai postId hoặc accountId");
+		else 
+			throw new AppException("Sai postId hoặc accountId");
 	}
 
 }
